Move auth logic into it's own file
